Refactoring so all content here works with Concord
Specifically changes so that the content here works as a concord import
and so that the content here works with the concord-terraform-processor-plugin.
A simple setup like this is now possible in a concord.yml file:

```
imports:
  - mvn:
      url: "mvn://ca.vanzyl.concord.terraform:concord-terraform:zip:1.0.1"
      dest: "."

configuration:
  arguments:
    aws:
      accessKey: "${crypto.exportAsString(projectInfo.orgName, 'xxx', null)}"
      secretKey: "${crypto.exportAsString(projectInfo.orgName, 'yyy', null)}"
      region: "us-east-2"
      vpcName: "default"
      instanceName: "concord-vm"
      instanceType: "m5.xlarge"
      instanceProfile: "ConcordCICDWorker"
      instanceKeypair: "automation"
      instanceVolumeSize: 10
      instanceTags:
        provisio.bucketPath: "s3://mybucket/releases"
        provisio.installerProfile: "cicd"
        provisio.applicationCoordinate: "ca.vanzyl:concord-agent:tar.gz:1.0.0"

flows:
  default:
    - call: ec2
      in:
        aws: ${aws}
```

All tests pass with the new setup, and corresponding tests are being
added to the concord-terraform-processor-plugin for each resource
type we currently handle.

Writing tests here with BATS, or writing tests in Java certainly has
trade offs one way or the other. Both can be made to work to assure the
Terraform content works correctly, but it's probably a team decision as
I think it would be hard to convince non-Java folks to want to write
tests in Java, and Java folks used to nice things like assertj likely
don't want to write tests in BASH but I think there might be value
in testing in both places. I'm going to try and flesh out the tests
in the concord-terraform-processor-plugin and here and see where
that leads.
works correctly